+++ libaegis/change/file/new.c
@@ -1,6 +1,6 @@
 /*
  *	aegis - project change supervisor
- *	Copyright (C) 1999 Peter Miller;
+ *	Copyright (C) 1999, 2002 Peter Miller;
  *	All rights reserved.
  *
  *	This program is free software; you can redistribute it and/or modify
@@ -36,7 +36,7 @@ change_file_new(cp, file_name)
 	fstate_src	*src_data_p;
 	type_ty		*type_p;
 
-	trace(("change_file_new(cp = %08lX)\n{\n"/*}*/, cp));
+	trace(("change_file_new(cp = %08lX)\n{\n"/*}*/, (long)cp));
 	fstate_data = change_fstate_get(cp);
 	assert(fstate_data->src);
 	src_data_p = fstate_src_list_type.list_parse(fstate_data->src, &type_p);
@@ -46,7 +46,7 @@ change_file_new(cp, file_name)
 	src_data->file_name = str_copy(file_name);
 	assert(cp->fstate_stp);
 	symtab_assign(cp->fstate_stp, file_name, src_data);
-	trace(("return %08lX;\n", src_data));
+	trace(("return %08lX;\n", (long)src_data));
 	trace((/*{*/"}\n"));
 	return src_data;
 }
